+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # No Comment Changelog
 
+## 0.4
+* Added confirmation dialogue after plugin activation. (Plugin will “soft”-activate, actual operations take place only after promted user confirmation.)
+* Added support for removing comment feeds from wp_head.
+
 ## 0.3
 * Added support for ping status closed.
 * Added option to prevent costly DB queries.